+++ src/game/Spells/SpellEntry.h
@@ -1067,7 +1067,7 @@ class SpellEntry
 
         int32 GetDuration() const;
         int32 GetMaxDuration() const;
-        int32 CalculateDuration(Unit const* caster = nullptr) const;
+        int32 CalculateDuration(WorldObject const* caster = nullptr) const;
         uint32 GetCastTime(Spell* spell = nullptr) const;
         uint32 GetCastTimeForBonus(DamageEffectType damagetype) const;
         uint16 GetAuraMaxTicks() const;
